requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
expressvpn-python==1.1
idna==2.10
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
user-agent==0.1.9
